fix(cli): runlock functional options, release-error logging, teardown lock, file splits
