+++ cli/internal/runlock/runlock_test.go
@@ -8,20 +8,18 @@ import (
 	"time"
 )
 
-// withShortWait shrinks the contention wait budget for the duration of a test
-// so a "second holder is blocked" assertion does not stall for the production
-// 5s budget. It restores the previous values on cleanup.
-func withShortWait(t *testing.T) {
-	t.Helper()
-	prevWait, prevRetry := WaitTimeout, RetryInterval
-	WaitTimeout = 150 * time.Millisecond
-	RetryInterval = 20 * time.Millisecond
-	t.Cleanup(func() {
-		WaitTimeout, RetryInterval = prevWait, prevRetry
-	})
+// shortWaitOpts shrink the contention wait budget so a "second holder is
+// blocked" assertion does not stall for the production 5s budget. They are
+// call-scoped (no shared package state), so every test can run in parallel.
+func shortWaitOpts() []Option {
+	return []Option{
+		WithWaitTimeout(150 * time.Millisecond),
+		WithRetryInterval(20 * time.Millisecond),
+	}
 }
 
 func TestAcquireRelease(t *testing.T) {
+	t.Parallel()
 	tests := []struct {
 		name    string
 		prelock bool // hold the lock before the subject Acquire
@@ -32,19 +30,19 @@ func TestAcquireRelease(t *testing.T) {
 	}
 	for _, tc := range tests {
 		t.Run(tc.name, func(t *testing.T) {
-			withShortWait(t)
+			t.Parallel()
 			dir := t.TempDir()
 			ctx := context.Background()
 
 			if tc.prelock {
-				held, err := Acquire(ctx, dir)
+				held, err := Acquire(ctx, dir, shortWaitOpts()...)
 				if err != nil {
 					t.Fatalf("prelock Acquire: %v", err)
 				}
 				t.Cleanup(func() { _ = held.Release() })
 			}
 
-			lock, err := Acquire(ctx, dir)
+			lock, err := Acquire(ctx, dir, shortWaitOpts()...)
 			if tc.wantErr != nil {
 				if !errors.Is(err, tc.wantErr) {
 					t.Fatalf("Acquire err = %v, want %v", err, tc.wantErr)
@@ -63,11 +61,11 @@ func TestAcquireRelease(t *testing.T) {
 }
 
 func TestReleaseAllowsReacquire(t *testing.T) {
-	withShortWait(t)
+	t.Parallel()
 	dir := t.TempDir()
 	ctx := context.Background()
 
-	first, err := Acquire(ctx, dir)
+	first, err := Acquire(ctx, dir, shortWaitOpts()...)
 	if err != nil {
 		t.Fatalf("first Acquire: %v", err)
 	}
@@ -75,21 +73,48 @@ func TestReleaseAllowsReacquire(t *testing.T) {
 		t.Fatalf("Release: %v", err)
 	}
 
-	second, err := Acquire(ctx, dir)
+	second, err := Acquire(ctx, dir, shortWaitOpts()...)
 	if err != nil {
 		t.Fatalf("re-Acquire after Release: %v", err)
 	}
 	t.Cleanup(func() { _ = second.Release() })
 }
 
+func TestCancelledContextIsNotErrLocked(t *testing.T) {
+	t.Parallel()
+	dir := t.TempDir()
+
+	// Hold the lock, then try to acquire it with an already-cancelled parent
+	// context: the caller aborted, so the error must NOT be ErrLocked.
+	held, err := Acquire(context.Background(), dir, shortWaitOpts()...)
+	if err != nil {
+		t.Fatalf("prelock Acquire: %v", err)
+	}
+	t.Cleanup(func() { _ = held.Release() })
+
+	ctx, cancel := context.WithCancel(context.Background())
+	cancel()
+	lock, err := Acquire(ctx, dir, shortWaitOpts()...)
+	if lock != nil {
+		t.Fatalf("Acquire returned a non-nil lock on cancelled context")
+	}
+	if errors.Is(err, ErrLocked) {
+		t.Fatalf("cancelled context surfaced ErrLocked, want a cancellation error: %v", err)
+	}
+	if !errors.Is(err, context.Canceled) {
+		t.Fatalf("Acquire err = %v, want context.Canceled", err)
+	}
+}
+
 func TestReleaseIsNilAndDoubleSafe(t *testing.T) {
+	t.Parallel()
 	var nilLock *Lock
 	if err := nilLock.Release(); err != nil {
 		t.Fatalf("nil Release: %v", err)
 	}
 
 	dir := t.TempDir()
-	lock, err := Acquire(context.Background(), dir)
+	lock, err := Acquire(context.Background(), dir, shortWaitOpts()...)
 	if err != nil {
 		t.Fatalf("Acquire: %v", err)
 	}
@@ -102,8 +127,9 @@ func TestReleaseIsNilAndDoubleSafe(t *testing.T) {
 }
 
 func TestLockFileLivesInDataDir(t *testing.T) {
+	t.Parallel()
 	dir := t.TempDir()
-	lock, err := Acquire(context.Background(), dir)
+	lock, err := Acquire(context.Background(), dir, shortWaitOpts()...)
 	if err != nil {
 		t.Fatalf("Acquire: %v", err)
 	}
